# compiler(amd64): generate trap exit code once