I've been looking at compiled code (amd64) and noticed out-of-bounds checks in `compileMemoryAccessCeilSetup` account for quite a bit of runtime, so I set out to try and optimize them a bit.

Basically on every instruction that might trap, we currently compile the trap and skip over it if a condition is met. This blows up compiled code significantly if there are a lot of instructions that might trap.

I kept pretty much the same logic, but cache a trap's address the first time it's compiled, and on future appearances directly jump into the previously compiled trap if the condition is *not* met.

This should not affect code generated for functions that either don't trap at all, or that trap only once (I tried not to make those worse). Functions that trap a lot get the benefit.

I'm making this a draft because I'm unsure if this the best a approach. I don't want to detour @mathetake from his efforts in overhauling the compiler (this is but a stopgap) but would appreciate comments.

Tested on my [`speedtest1`](https://github.com/ncruces/go-sqlite3/tree/main/vfs/tests/speedtest1) SQLite benchmark:
```sh
speedtest1 % (cd ~/wazero; git switch main)
speedtest1 % go test -bench=. -benchtime=50x -count=10 --memdb | tee before.txt
speedtest1 % (cd ~/wazero; git switch traps)
speedtest1 % go test -bench=. -benchtime=50x -count=10 --memdb | tee trap.txt
```

Compiled code size goes down from 10.1MiB to 8.1MiB (-20%).
Benchstat results:
```sh
goos: darwin
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/vfs/tests/speedtest1
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
               │ before.txt  │              trap.txt               │
               │   sec/op    │   sec/op     vs base                │
_speedtest1-12   87.27m ± 3%   75.95m ± 1%  -12.96% (p=0.000 n=10)
```

@codefromthecrypt any ideas what other benchmarks I should try?