@ncruces 2023-06-05
One alternative I considered, but which I didn't implement, is to compile each trap once per module, regardless of it being used.

I tried doing it in the function preamble, but this makes functions that don't trap worse, which I wanted to avoid. Results for functions that _do_ trap are similar, slightly _better_ even.

But since doing it once per module requires `engine.go` changes, and I don't have an `arm64`, I did this for a start.