@ncruces 2023-06-05
The force push is just a rebase, to make sure the tests pass after @evacchi fix.

I tried to extend this idea to the "trapping float to int conversions" and I thought it _might_ be clearer if the condition was kept inverted (because there's always trapping and non-trapping branches), that leads to this patch instead:
https://github.com/tetratelabs/wazero/compare/main...ncruces:wazero:trap2

I honestly don't know which one reads better.

Float to int conversions aren't common enough to make a performance difference, so we may as well ignore them: 95% of the benefit is from memory access, and the other 5% from table access for indirect calls.