@mathetake 2023-06-05
let's do the same to trapping int to float as well for consistency