@ncruces 2023-06-05
- added most trapping int to float
- renamed variables in `compileTrapFromNativeCode` for readability
- fixed some comments for clarity
- tested again to confirm no regressions

Benchstat results:
```
goos: darwin
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/vfs/tests/speedtest1
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
               │ before.txt  │              trap.txt               │
               │   sec/op    │   sec/op     vs base                │
_speedtest1-12   88.91m ± 2%   75.64m ± 2%  -14.92% (p=0.000 n=10)
```